+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+# [0.2.0](/compare/v0.1.0...v0.2.0) (2025-12-13)
+
+
+### Features
+
+* **root:** added button ui component + fix size root layout 75b3d0d
+* **root:** added section - layout - buttons - pill e1e7208
+* **root:** created components - DS 3b1f354
+
 # 0.1.0 (2025-12-12)